requirements.txt
pyyaml
schema
pymongo
asyncio
python-telegram-bot
pytest
datetime
pulp
numpy
fastapi
uvicorn
